requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.4
celery==5.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==6.0.2
django-appconf==1.2.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-geoip==0.5.2.1
django-ip-geolocation==1.6.1
django-ipware==7.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
idna==3.11
kombu==5.6.2
packaging==25.0
promise==2.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-ipware==3.0.0
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
